Fix secu issues

+++ src/features/perform_privacy_operation/cmd_perform_privacy_operation.c
@@ -125,5 +125,5 @@ end:
     explicit_bzero(privateKeyDataSwapped, sizeof(privateKeyDataSwapped));
     explicit_bzero(&privateKey, sizeof(privateKey));
     explicit_bzero(privateKeyData, sizeof(privateKeyData));
-    return error;
+    return (uint16_t) error;
 }
